backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
python-jose==3.3.0
python-dotenv==1.0.0
dlib==19.24.2
face-recognition==1.3.0
rembg==2.0.50
scikit-image==0.21.0 